requirements.txt
# TableTalk — Python dependencies
# Install with: pip install -r requirements.txt

# Core audio processing
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0

# Data manipulation & ML
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep learning (optional — needed only for neural net experiments)
# torch>=2.0.0
# torchaudio>=2.0.0

# Transcription
openai-whisper>=20230918
jiwer>=3.0.3          # WER / CER metrics

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0